\name{interactivity}
\docType{package}
\title{
Toggle interactivity in R
}
\description{
This very small package provides a function to switch between interactive and non-interactive mode in R.
}
\details{
\tabular{ll}{
Package: \tab interactivity\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-08-11\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
Jeffrey Horner, Jeroen Ooms
Maintainer: Jeroen Ooms <jeroenooms@gmail.com>

}
\references{
Also see function interactive (base)
}

\keyword{ interactive }
\seealso{

\code{\link[base:interactive]{interactive}} 
}
\examples{
setInteractive(FALSE);
interactive();
}
